split cuda vs others compilation into two test files.
